+++ utils/utils.h
@@ -7,8 +7,8 @@ void save_palette(int16_t *paletteBuffer);
 int16_t st_Getrez(void);
 
 void reset(void);
-int32_t get200hz(void);
-int32_t st_Supexec(int32_t atari_hardware_addr());
+long get200hz(void);
+long st_Supexec(long atari_hardware_addr());
 
 int16_t st_Dgetpath(char* current_path, int16_t drive_num);
 int16_t st_Dgetdrv();
@@ -23,7 +23,7 @@ int16_t st_form_alert_choice(int16_t this_form_icon, char* this_form_message, ch
 void st_clear_char_array(char* carray);
 char *basename(char const *path);
 
-void *mem_alloc(int32_t size);
+void *mem_alloc(long size);
 void *mem_free(void *ptr);
 
 MFDB* mfdb_alloc_bpp( int8_t* buffer, int16_t width, int16_t height, int16_t bpp);
